server/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-js-asset==2.1.0
django-mptt==0.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
googlemaps==4.10.0
idna==3.4
psycopg2==2.9.6
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
stripe==5.5.0
urllib3==2.0.4
